// src/ui/panels/types.ts
export type ActiveTool =
	| 'transform'
	| 'select'
	| 'wand'
	| 'crop'
	| 'eyedropper'
	| 'retouch'
	| 'brush'
	| 'history'
	| 'clone'
	| 'eraser'
	| 'fill'
	| 'gradient'
	| 'tone'
	| 'text'
	| 'shape'
	| 'path'
	| 'hand'
	| 'zoom';